Half-precision float support

+++ tce/src/base/osal/OperationPimpl.cc
@@ -852,7 +852,7 @@ llvmOperandType(Operand::OperandType type) {
     case Operand::UINT_WORD:
         return "i32";
     case Operand::HALF_FLOAT_WORD:
-        return "i32";
+        return "f16";
     case Operand::FLOAT_WORD:
         return "f32";
     case Operand::DOUBLE_WORD:
